small refactor + BeginTest component
