+++ src/components/helpers/DownloadLogs.vue
